switch to vue-codemirror6, fix onCmCodeChange, make splitter work
