+++ docs/.vitepress/vite-env.d.ts
@@ -1,4 +1,4 @@
 declare module '*.tex?raw' {
-  const content: string
-  export default content
+    const content: string
+    export default content
 }
