+++ package.json
@@ -91,7 +91,6 @@
     "@codemirror/autocomplete": "6.x",
     "@codemirror/commands": "6.x",
     "@codemirror/language": "6.x",
-    "@codemirror/legacy-modes": "6.x",
     "@codemirror/lint": "6.x",
     "@codemirror/search": "6.x",
     "@codemirror/state": "6.x",
@@ -99,8 +98,7 @@
     "@codemirror/theme-one-dark": "6.x",
     "codemirror-lang-latex": "0.4.x",
     "vue-tsc": "3.3.x",
-    "vue-codemirror": "^6.1.1",
-    "stylus": "0.59.x",
+    "vue-codemirror6": "1.5.x",
     "vite": "8.0.x",
     "vite-plugin-node-polyfills": "0.28.x",
     "vite-plugin-checker": "0.14.x",
